Extract static methods for evaluating individual points in both 2d and 3d cases.
