+++ RenderItems/Graph3dCurveItem.cs
@@ -144,28 +144,7 @@ namespace MetaphysicsIndustries.Ligra.RenderItems
                 env.SetVariable(varName, new Literal(x));
 
                 var vv = expr.Eval(env);
-                if (!vv.IsConcrete)
-                    // EvaluationException ?
-                    throw new OperandException(
-                        "Value is not concrete");
-
-                Vector3 pt;
-                if (!vv.IsVector(null))
-                    throw new OperandException(
-                        "Value is not a 3-vector");
-                if (vv.GetVectorLength(null) != 3)
-                    // EvaluationException ?
-                    throw new OperandException(
-                        "Value is not a 3-vector");
-                var vvv = vv.ToVector();
-                // TODO: check for NaN
-                // TODO: ensure components of vvv are scalars
-                pt = new Vector3(
-                    vvv[0].ToNumber().Value,
-                    vvv[1].ToNumber().Value,
-                    vvv[2].ToNumber().Value);
-
-                points[i] = pt;
+                points[i] = GraphItemUtil.EvaluatePoint3d(vv);
             }
         }
